Namespace all files under uaa/
With the way things are currently layed out, it's not possible to
install this and bosh-cli alongside each other.  This is becaise both
have a `cli/config.rb` file, and therefore when bosh-cli does a `require
'cli/config'` it gets the uaa version instead of its own.

Fix this by namespacing all the files under `uaa/` to avoid this
collision. This also makes this gem better adhere to the recommended
rubygems patterns - http://guides.rubygems.org/patterns/#loading-code