Support ca_file for http client in curl mode
With curl simulation mode the parameters from curl needs to be passed
The parameters from uaac target xxx are not used here.
Therefore, ca cert was missing. -b for --skip-ssl-validation is available
For uaac target --ca-cert there is was no setting.
Added -C or --cacert because these are the parameters from curl